Add day 11 for AOC 2023

+++ 2023/include/day6.h
@@ -17,7 +17,7 @@ namespace aoc
         j++;
       }
       auto ways = time - j - j + 1;
-      count *= ways;
+      count *= (int) ways;
     }
     return count;
   }
